cgda/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-dotenv==1.2.1
python-jose[cryptography]==3.3.0
passlib==1.7.4
python-multipart==0.0.19
SQLAlchemy==2.0.36
pydantic==2.10.3
requests==2.32.3
reportlab==4.2.5
pandas==2.2.2
openpyxl==3.1.5


